Pick Ups
There are now pick-up items that trigger specific events to the player. These include the growing and shrinking potions, and easter egg items that will be hidden around the levels in hard to reach places. These items have been placed all around the maps and in the relevant places that they are needed in each level